requirements.txt
# QM Final Project (Python 3.10+)
pandas>=2.0
numpy>=1.24
matplotlib>=3.7
seaborn>=0.13
yfinance>=0.2.40
statsmodels>=0.14
scikit-learn>=1.3
scipy>=1.11
jupyter>=1.0
nbformat>=5.9
nbclient>=0.8
python-pptx>=0.6.21
